feat(termination) - time off component (#96)
* time off component

* simplify timeoff

* use render prop

* fix timeoff

* tests

* change API

* apply changes

* remove example

---------
