+++ src/flows/Termination/tests/fixtures.ts
@@ -1,0 +1,3 @@
+export const timeoff = {
+  data: { total_count: 0, current_page: 1, total_pages: 0, timeoffs: [] },
+};
